Moved files into private sub-directory
