+++ include/siddiqsoft/private/openssl_helpers.hpp
