+++ include/siddiqsoft/private/restcl_definitions.hpp
