+++ include/siddiqsoft/private/restcl_win.hpp
@@ -1,42 +1,19 @@
-/*
-    restcl : WinHTTP implementation for Windows
-
-    BSD 3-Clause License
-
-    Copyright (c) 2021, Siddiq Software LLC
-    All rights reserved.
-
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions are met:
-
-    1. Redistributions of source code must retain the above copyright notice, this
-       list of conditions and the following disclaimer.
-
-    2. Redistributions in binary form must reproduce the above copyright notice,
-       this list of conditions and the following disclaimer in the documentation
-       and/or other materials provided with the distribution.
-
-    3. Neither the name of the copyright holder nor the names of its
-       contributors may be used to endorse or promote products derived from
-       this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-    AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-    IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-    FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-    DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-    SERVICES; LOSS OF USE, d_, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-    CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-    OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-    OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+/**
+ * @file restcl_win.hpp
+ * @author Abdulkareem Siddiq (github@siddiqsoft.com)
+ * @brief WinHTTP based implementation of the basic_restclient
+ * @version
+ * @date 2024-12-24
+ *
+ * @copyright Copyright (c) 2024 Siddiq Software
+ *
  */
 
 #pragma once
 #if defined(WIN32) || defined (WIN64) || defined (_WIN32) || defined (_WIN64)
 
-#ifndef RESTCL_WINHTTP_HPP
-#define RESTCL_WINHTTP_HPP
+#ifndef RESTCL_WIN_HPP
+#define RESTCL_WIN_HPP
 
 #include <iostream>
 #include <chrono>
@@ -63,7 +40,7 @@
 
 
 #include "nlohmann/json.hpp"
-#include "restcl.hpp"
+#include "../restcl.hpp"
 
 #include "siddiqsoft/string2map.hpp"
 #include "siddiqsoft/acw32h.hpp"
