+++ tests/test_winhttp.cpp
@@ -40,7 +40,6 @@
 
 #include "nlohmann/json.hpp"
 #include "../include/siddiqsoft/restcl.hpp"
-#include "../include/siddiqsoft/restcl_winhttp.hpp"
 
 
 namespace siddiqsoft
